- Added: Refactor general: eliminado codigo comentado obsoleto, variables e imports sin usar, etc. Eliminados muchos warnings, de 645 se ha pasado a solo 25.

+++ app/src/main/java/com/rds/rodalies/alarmas/NuevaAlarmaNotificacion.java
@@ -72,8 +72,8 @@ public class NuevaAlarmaNotificacion extends FragmentActivity {
         if(domingo.isChecked() || lunes.isChecked() || martes.isChecked() || miercoles.isChecked() || jueves.isChecked() ||
                 viernes.isChecked() || sabado.isChecked()){
 
-            alarmDetails.timeMinute = timePicker1.getCurrentMinute().intValue();
-            alarmDetails.timeHour = timePicker1.getCurrentHour().intValue();
+            alarmDetails.timeMinute = timePicker1.getCurrentMinute();
+            alarmDetails.timeHour = timePicker1.getCurrentHour();
 
             alarmDetails.setRepeatingDay(AlarmModel.SUNDAY, domingo.isChecked());
             alarmDetails.setRepeatingDay(AlarmModel.MONDAY, lunes.isChecked());
